+++ src/main/java/org/ojalgo/matrix/BasicMatrix.java
@@ -29,8 +29,6 @@ import java.util.function.Supplier;
 import org.ojalgo.ProgrammingError;
 import org.ojalgo.RecoverableCondition;
 import org.ojalgo.algebra.NormedVectorSpace;
-import org.ojalgo.algebra.Operation;
-import org.ojalgo.algebra.ScalarOperation;
 import org.ojalgo.function.aggregator.Aggregator;
 import org.ojalgo.function.aggregator.AggregatorFunction;
 import org.ojalgo.function.constant.PrimitiveMath;
@@ -66,21 +64,11 @@ import org.ojalgo.type.context.NumberContext;
  *
  * @author apete
  */
-public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix<N, M>> implements NormedVectorSpace<M, N>, Operation.Subtraction<M>,
-        Operation.Multiplication<M>, ScalarOperation.Addition<M, N>, ScalarOperation.Division<M, N>, ScalarOperation.Subtraction<M, N>, Access2D<N>,
-        Access2D.Elements, Access2D.Aggregatable<N>, Structure2D.ReducibleTo1D<M>, NumberContext.Enforceable<M>, Access2D.Collectable<N, PhysicalStore<N>> {
+public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix<N, M>>
+        implements Matrix2D<N, M>, Access2D.Elements, Structure2D.ReducibleTo1D<M>, NumberContext.Enforceable<M>, Access2D.Collectable<N, PhysicalStore<N>> {
 
     public interface LogicalBuilder<N extends Comparable<N>, M extends BasicMatrix<N, M>>
             extends Structure2D.Logical<M, BasicMatrix.LogicalBuilder<N, M>>, Access2D.Collectable<N, PhysicalStore<N>> {
-
-        default M build() {
-            return this.get();
-        }
-
-    }
-
-    public interface PhysicalReceiver<N extends Comparable<N>, M extends BasicMatrix<N, M>>
-            extends Mutate2D.ModifiableReceiver<N>, Supplier<M>, Access2D.Collectable<N, PhysicalStore<N>> {
 
         default M build() {
             return this.get();
@@ -210,7 +198,7 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
     /**
      * @return A fully mutable matrix builder with the elements initially set to a copy of this matrix.
      */
-    public abstract BasicMatrix.PhysicalReceiver<N, M> copy();
+    public abstract <R extends Mutate2D.ModifiableReceiver<N> & Supplier<M>> R copy();
 
     public long count() {
         return myStore.count();
@@ -766,7 +754,7 @@ public abstract class BasicMatrix<N extends Comparable<N>, M extends BasicMatrix
 
     abstract SingularValue<N> getDecompositionSingularValue(Structure2D typical);
 
-    abstract MatrixFactory<N, M, ? extends LogicalBuilder<N, M>, ? extends PhysicalReceiver<N, M>, ? extends PhysicalReceiver<N, M>> getFactory();
+    abstract MatrixFactory<N, M, ? extends LogicalBuilder<N, M>, ?, ?> getFactory();
 
     final MatrixStore<N> getStore() {
         return myStore;