+++ src/main/java/org/ojalgo/matrix/store/Primitive32Store.java
@@ -763,23 +763,23 @@ public final class Primitive32Store extends Primitive32Array implements Physical
     }
 
     public TransformableRegion<Double> regionByColumns(final int... columns) {
-        return new TransformableRegion.ColumnsRegion<>(this, multiplyBoth, columns);
+        return new Subregion2D.ColumnsRegion<>(this, multiplyBoth, columns);
     }
 
     public TransformableRegion<Double> regionByLimits(final int rowLimit, final int columnLimit) {
-        return new TransformableRegion.LimitRegion<>(this, multiplyBoth, rowLimit, columnLimit);
+        return new Subregion2D.LimitRegion<>(this, multiplyBoth, rowLimit, columnLimit);
     }
 
     public TransformableRegion<Double> regionByOffsets(final int rowOffset, final int columnOffset) {
-        return new TransformableRegion.OffsetRegion<>(this, multiplyBoth, rowOffset, columnOffset);
+        return new Subregion2D.OffsetRegion<>(this, multiplyBoth, rowOffset, columnOffset);
     }
 
     public TransformableRegion<Double> regionByRows(final int... rows) {
-        return new TransformableRegion.RowsRegion<>(this, multiplyBoth, rows);
+        return new Subregion2D.RowsRegion<>(this, multiplyBoth, rows);
     }
 
     public TransformableRegion<Double> regionByTransposing() {
-        return new TransformableRegion.TransposedRegion<>(this, multiplyBoth);
+        return new Subregion2D.TransposedRegion<>(this, multiplyBoth);
     }
 
     public RowView<Double> rows() {